cambios clase 14/11/2024

+++ AD/HospitalCRUDSpring/src/main/java/com/hospitalcrud/dao/respositories/spring/SpringPatientsRepository.java
@@ -5,6 +5,7 @@ import com.hospitalcrud.dao.model.Patient;
 import com.hospitalcrud.dao.respositories.*;
 import com.hospitalcrud.dao.utilities.Constantes;
 import com.hospitalcrud.dao.utilities.SQLQueries;
+import com.hospitalcrud.domain.error.FOREIGN_KEY_ERROR;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
 import org.springframework.jdbc.core.simple.JdbcClient;
@@ -13,6 +14,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Objects;
 
@@ -62,12 +64,14 @@ public class SpringPatientsRepository implements PatientRepository {
 
     @Override
     @Transactional
-    public boolean delete(int patientId) {
-        medicationsRepository.deletePatientMedications(patientId);
-        medicalRecordsRepository.deletePatientMedicalRecords(patientId);
+    public boolean delete(int patientId, boolean comfirmation) {
+        if (comfirmation) {
+            medicationsRepository.deletePatientMedications(patientId);
+            medicalRecordsRepository.deletePatientMedicalRecords(patientId);
+        }
         credentialRepository.delete(patientId);
         paymentsRepository.delete(patientId);
-        jdbcClient.sql(SQLQueries.DELETE_PATIENT).param("id",patientId);
+        jdbcClient.sql(SQLQueries.DELETE_PATIENT).param("id", patientId).update();
         return false;
     }
 }
